refactor: Extract cni manager as the public part
